backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
coverage==7.13.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
